+++ firmware/src/worker.rs
@@ -293,6 +293,7 @@ impl<'a> Worker<'a> {
                 presence,
                 presence,
                 presence,
+                presence,
                 platform,
                 serial_id,
                 serial_hash,